requirements.txt
Django==5.2.1
dj-database-url==2.2.0
python-dotenv==1.0.1
psycopg[binary]==3.2.10
gunicorn==23.0.0
whitenoise==6.9.0
cryptography==45.0.1
